Extract `MockTimeSource` into its own class.
