+++ src/clocks/index.js
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from '#x/IntfTimeSource';
+export * from '#x/MockTimeSource';
 export * from '#x/StdTimeSource';
 export * from '#x/WallClock';
